feat: seprate listing of every student

+++ src/pages/Listing/index.js
@@ -1,12 +1,20 @@
 import React, { useEffect } from "react";
 import { useState } from "react";
 import Layout from "../../utils/Layout";
-import { Navigate, useNavigate } from "react-router-dom";
+import { Navigate, useLocation, useNavigate } from "react-router-dom";
 import { Button, OutlinedButton } from "../../components/Shared";
 import { getUser } from "../../utils/utils";
 const Listing = () => {
 
+
+    const loginUsers = JSON.parse(localStorage.getItem("loginUser"));
+    const loggedInUser = loginUsers.find(user => user.isLogin === true)
+
+    console.log(loggedInUser)
+
+    const { state } = useLocation();
     const navigate = useNavigate();
+    console.log("id", state)
 
     //set the array of users
     const [userData, setUserData] = useState(getUser());
@@ -44,6 +52,7 @@ const Listing = () => {
         navigate("/add-new-student")
     }
 
+
     return (
         <Layout>
             <div className='w-full'>
@@ -80,20 +89,22 @@ const Listing = () => {
                         <tbody>
                             {
                                 userData.map((user, index) => {
-                                    return (
-                                        <tr key={index} className='py-2 text-center'>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{index + 1}</td>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{user.username}</td>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{user.email}</td>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{user.age}</td>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{user.address}</td>
-                                            <td className='border border-primaryColor px-6 text-nowrap'>{user.gender}</td>
-                                            <td className='flex justify-center items-center gap-2  px-6 text-textColor border border-primaryColor'>
-                                                <button className='bg-dangerColor w-14 h-8 rounded-sm' onClick={() => deleteUser(index)}>Delete</button>
-                                                <button className='bg-successColor w-14 h-8 rounded-sm' onClick={() => editUser(user)}>Edit</button>
-                                            </td>
-                                        </tr>
-                                    )
+                                    if (user.parentId === loggedInUser.id) {
+                                        return (
+                                            <tr key={index} className='py-2 text-center'>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{index + 1}</td>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{user.username}</td>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{user.email}</td>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{user.age}</td>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{user.address}</td>
+                                                <td className='border border-primaryColor px-6 text-nowrap'>{user.gender}</td>
+                                                <td className='flex justify-center items-center gap-2  px-6 text-textColor border border-primaryColor'>
+                                                    <button className='bg-dangerColor w-14 h-8 rounded-sm' onClick={() => deleteUser(index)}>Delete</button>
+                                                    <button className='bg-successColor w-14 h-8 rounded-sm' onClick={() => editUser(user)}>Edit</button>
+                                                </td>
+                                            </tr>
+                                        )
+                                    }
                                 })
                             }
                         </tbody>
